ft-(user-login): Users should be able to Login to Barefoot Nomad
