+++ src/App.js
@@ -3,7 +3,6 @@ import './styles/index.scss'
 import Nav from './dammyNav'
 import { BrowserRouter as Router } from 'react-router-dom'
 import AllRoutes from './routes';
-import Login from './components/login';
 
 function App() {
   return (
@@ -11,7 +10,6 @@ function App() {
       <Router>
         <Nav />
         <AllRoutes />
-        <Login />
       </Router>
     </div>
   )
